Rework of the PWM module for both linux and nuttx v7.
1. API unification (period in seconds, duty cycle in units)
2. Feature parity (setPeriod and setFrequency supported on both platforms)
3. State transition closure (period and duty cycle may be passed an any time)
4. Parameter verification (NaNs/inf handling, clamp to range)
5. [linux] Factor device path computing out to a function

IoT.js-DCO-1.0-Signed-off-by: Tomasz Wozniak t.wozniak@samsung.com

+++ src/module/iotjs_module_pwm.h
@@ -43,8 +43,8 @@ typedef enum {
 
 typedef struct {
   int32_t pin;
-  int32_t duty_cycle;
-  int32_t period;
+  double duty_cycle;
+  double period;
   bool enable;
 
   PwmError result;
